+++ src/Core/Dfe.Complete.Application/Notes/Commands/CreateNoteCommand.cs
@@ -3,7 +3,7 @@ using Dfe.Complete.Application.Notes.Interfaces;
 using Dfe.Complete.Domain.Entities;
 using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Domain.ValueObjects;
-using DfE.CoreLibs.Utilities.Extensions;
+using GovUK.Dfe.CoreLibs.Utilities.Extensions;
 using MediatR;
 using Microsoft.Extensions.Logging;