+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/ProjectsController/TaskDataControllerTests.cs
@@ -4,8 +4,8 @@ using Dfe.Complete.Client.Contracts;
 using Dfe.Complete.Domain.Entities;
 using Dfe.Complete.Infrastructure.Database;
 using Dfe.Complete.Tests.Common.Constants;
-using DfE.CoreLibs.Testing.AutoFixture.Attributes;
-using DfE.CoreLibs.Testing.Mocks.WebApplicationFactory;
+using GovUK.Dfe.CoreLibs.Testing.AutoFixture.Attributes;
+using GovUK.Dfe.CoreLibs.Testing.Mocks.WebApplicationFactory;
 using Microsoft.EntityFrameworkCore;
 using System.Security.Claims;
 
@@ -283,7 +283,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
             command.SignedSecretaryState = true;
 
             // Act
-            await tasksDataClient.UpdateSupplementalFundingAgreementTaskAsync(command, default);
+            await tasksDataClient.UpdateSupplementalFundingAgreementTaskAsync(command);
 
             // Assert
             dbContext.ChangeTracker.Clear();
@@ -320,7 +320,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
             command.Received = true;
 
             // Act
-            await tasksDataClient.UpdateSupplementalFundingAgreementTaskAsync(command, default);
+            await tasksDataClient.UpdateSupplementalFundingAgreementTaskAsync(command);
 
             // Assert
             dbContext.ChangeTracker.Clear();
@@ -508,6 +508,80 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
             Assert.Null(existingTaskData.ProposedCapacityOfTheAcademyReceptionToSixYears);
             Assert.Null(existingTaskData.ProposedCapacityOfTheAcademyTwelveOrAboveYears);
             Assert.Null(existingTaskData.ProposedCapacityOfTheAcademySevenToElevenYears);
+        }
+
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateDeclarationOfExpenditureCertificateTaskAsync_ShouldUpdate_ConversionTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateDeclarationOfExpenditureCertificateTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
+            command.ProjectType = ProjectType.Conversion;
+            command.CheckCertificate = true;
+            command.Saved = false;
+            command.NotApplicable = false;
+            command.DateReceived = new DateTime(2025, 1, 1);
+
+            // Act
+            await tasksDataClient.UpdateDeclarationOfExpenditureCertificateTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.False(existingTaskData.ReceiveGrantPaymentCertificateNotApplicable);
+            Assert.True(existingTaskData.ReceiveGrantPaymentCertificateCheckCertificate);
+            Assert.NotNull(existingTaskData.ReceiveGrantPaymentCertificateDateReceived);
+            Assert.False(existingTaskData.ReceiveGrantPaymentCertificateSaveCertificate); 
+        }
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateDeclarationOfExpenditureCertificateTaskAsync_ShouldUpdate_TransferTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateDeclarationOfExpenditureCertificateTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<TransferTasksData>();
+            dbContext.TransferTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
+            command.ProjectType = ProjectType.Transfer;
+            command.CheckCertificate = true;
+            command.Saved = false;
+            command.NotApplicable = false;
+            command.DateReceived = new DateTime(2025, 1, 1);
+
+            // Act
+            await tasksDataClient.UpdateDeclarationOfExpenditureCertificateTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.TransferTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.NotNull(existingTaskData);
+            Assert.False(existingTaskData.DeclarationOfExpenditureCertificateSaved);
+            Assert.True(existingTaskData.DeclarationOfExpenditureCertificateCorrect);
+            Assert.NotNull(existingTaskData.DeclarationOfExpenditureCertificateDateReceived);
+            Assert.False(existingTaskData.DeclarationOfExpenditureCertificateNotApplicable);
         }
     }
 }