+++ src/Tests/Dfe.Complete.Application.Tests/QueryHandlers/Project/GetProjectGroupByIdQueryHandlerTests.cs
@@ -1,8 +1,8 @@
 ﻿using AutoFixture.Xunit2;
-using DfE.CoreLibs.Testing.AutoFixture.Attributes;
+using GovUK.Dfe.CoreLibs.Testing.AutoFixture.Attributes;
 using Dfe.Complete.Domain.Interfaces.Repositories;
 using NSubstitute;
-using DfE.CoreLibs.Testing.AutoFixture.Customizations;
+using GovUK.Dfe.CoreLibs.Testing.AutoFixture.Customizations;
 using System.Linq.Expressions;
 using Dfe.Complete.Domain.ValueObjects;
 using AutoMapper;