+++ src/Tests/Dfe.Complete.Application.Tests/Services/CsvExport/Builders/FormAMatTests.cs
@@ -1,7 +1,7 @@
 ﻿using Dfe.Complete.Application.Services.CsvExport.Builders;
 using Dfe.Complete.Domain.Entities;
 using Dfe.Complete.Tests.Common.Customizations.Models;
-using DfE.CoreLibs.Testing.AutoFixture.Attributes;
+using GovUK.Dfe.CoreLibs.Testing.AutoFixture.Attributes;
 
 namespace Dfe.Complete.Application.Tests.Services.CsvExport.Builders
 {